+++ src/environments/environment.ts
@@ -14,4 +14,5 @@ export const environment = {
   PROXY_API_URL: () => `${environment.TC_ROOT_API_URL}/proxy`,
   PROPERTIES_API_URL: () => `${environment.TC_ROOT_API_URL}/properties`,
   AUDIT_API_URL: () => `${environment.TC_ROOT_API_URL}/audit`,
+  DASHBOARD_API_URL: () => `${environment.TC_ROOT_API_URL}/dashboard`,
 };